mod: 게시글 조회핸들러와 조회수 증가 핸들러 통합
